Refactor: URLSession 관련 메소드는 네트워크 관련이므로 	  네트워크 클래스를 만들어서 옮김
