+++ PhotosApp/PhotosApp/Utils/Util.swift
@@ -10,21 +10,6 @@ import Foundation
 
 class Util {
     
-    static func excuteURLSession(from urlString: String, completion: @escaping (Data?) -> ()) {
-        guard let url = URL(string: urlString)
-            else {
-                return
-        }
-        
-        URLSession.shared.dataTask(with: url) { data, response, error in
-            if let error = error {
-                print(error.localizedDescription)
-            } else if let data = data {
-                completion(data)
-            }
-        }.resume()
-    }
-    
     static func decodeJSONData<T>(from urlString: String, type: T.Type, completion: @escaping (T?) -> ())
         where T: Decodable {
             excuteURLSession(from: urlString) { (data) in
